Specialize WalIngest on PostgreSQL version
The current code assumes that most of this functionality is version-independent,
which is only true up to v16 - PostgreSQL 17 has a new field in CheckPoint that
we need to keep track of.

This basically removes the file-level dependency on v14, and replaces it with
switches that load the correct version dependencies where required.