storcon: forward request to leader when stepped down
More specifically we forward when:
1. Request is not one of ["/control/v1/step_down", "/status", "/ready", "/metrics"]
2. Current instance is in [`LeadershipStatus::SteppedDown`] state
3. There is a leader in the database to forward to
4. Leader from step (3) is not the current instance

Why forward?
It turns out that we can't rely on external orchestration to promptly route trafic to the
new leader. This is downtime inducing. Forwarding provides a safe way out.

Why is it safe?
If a storcon instance is persisted in the database, then we know that it is the current leader.
There's one exception: time between handling step-down request and the new leader updating the
database.

Let's treat the happy case first. The stepped down node does not produce any side effects,
since all request handling happens on the leader.

As for the edge case, we are guaranteed to always have a maximum of two running instances.
Hence, if we are in the edge case scenario the leader persisted in the database is the
stepped down instance that received the request. Condition (4) above covers this scenario.